backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-multipart==0.0.20
cryptography==41.0.7
supabase==2.12.0
python-dotenv==1.0.1
email-validator==2.1.1
huggingface-hub
llama-cpp-python
